Fixed lookup of team player name for identifying which board to print.
Previously we checked the "user" member of a PlayerPresenter,
which doesn't exist. Now we check the "username" member,
which does. This also means that comparison is done with the
username rather than with whatever primary key a user object
has; this is good practice because someone might swap in
a new user class. See
https://docs.djangoproject.com/en/dev/ref/contrib/auth/#django.contrib.auth.models.User.get_username